// src/enums/Action.ts
export enum Action {
  CREATE_SUBACCOUNT = 'CREATE_SUBACCOUNT',
  GET_SUBACCOUNT = 'GET_SUBACCOUNT',
  LIST_SUBACCOUNT = 'LIST_SUBACCOUNT',

  CRAETE_BILL = 'CREATE_BILL',
  GET_BILL = 'GET_BILL',
  LIST_BILL = 'LIST_BILL',
  SIMULATE_BILL = 'SIMULATE_BILL',
  CANCEL_BILL = 'CANCEL_BILL',

  CREATE_PAYMENT = 'CREATE_PAYMENT',
  GET_PAYMENT = 'GET_PAYMENT',
  LIST_PAYMENT = 'LIST_PAYMENT',
  DELETE_PAYMENT = 'DELETE_PAYMENT',
  RESTORE_PAYMENT = 'RESTORE_PAYMENT',
  UPDATE_PAYMENT = 'UPDATE_PAYMENT',
  REFUND_PAYMENT = 'REFUND_PAYMENT',
  GET_IDENTIFICATION_PAYMENT = 'GET_IDENTIFICATION_PAYMENT',
  GET_PIX_QRCODE_PAYMENT = 'GET_PIX_QRCODE_PAYMENT',
  RECEIVED_IN_CASH_PAYMENT = 'RECEIVED_IN_CASH_PAYMENT',
  UNDO_RECEIVED_IN_CASH_PAYMENT = 'UNDO_RECEIVED_IN_CASH_PAYMENT',
  GET_LIMITS_PAYMENT = 'LIMITS_PAYMENT',

  CREATE_CUSTOMER = 'CREATE_CUSTOMER',
  GET_CUSTOMER = 'GET_CUSTOMER',
  LIST_CUSTOMER = 'LIST_CUSTOMER',
  UPDATE_CUSTOMER = 'UPDATE_CUSTOMER',
  DELETE_CUSTOMER = 'DELETE_CUSTOMER',
  RESTORE_CUSTOMER = 'RESTORE_CUSTOMER',

  LIST_INSTALLMENT = 'LIST_INSTALLMENT',
  GET_INSTALLMENT = 'GET_INSTALLMENT',
  GET_PAYMENT_BOOK_INSTALLMENT = 'GET_PAYMENT_BOOK_INSTALLMENT',
  DELETE_INSTALLMENT = 'DELETE_INSTALLMENT',
  REFUND_INSTALLMENT = 'REFUND_INSTALLMENT',

  CREATE_INVOICE = 'CREATE_INVOICE',
  GET_INVOICE = 'GET_INVOICE',
  LIST_INVOICE = 'LIST_INVOICE',
  CANCEL_INVOICE = 'CANCEL_INVOICE',
  AUTHORIZE_INVOICE = 'AUTHORIZE_INVOICE',
  UPDATE_INVOICE = 'UPDATE_INVOICE',

  GET_COMMERCIAL_INFO_ACCOUNT = 'GET_COMMERCIAL_INFO_ACCOUNT',
  UPDATE_COMMERCIAL_INFO_ACCOUNT = 'UPDATE_COMMERCIAL_INFO_ACCOUNT',
  UPDATE_PAYMENT_CHECKOUT_CONFIG_ACCOUNT = 'UPDATE_PAYMENT_CHECKOUT_CONFIG_ACCOUNT',
  GET_PAYMENT_CHECKOUT_CONFIG_ACCOUNT = 'GET_PAYMENT_CHECKOUT_CONFIG_ACCOUNT',
  GET_ACCOUNT_NUMBER_ACCOUNT = 'GET_ACCOUNT_NUMBER_ACCOUNT',
  GET_FEES_ACCOUNT = 'GET_FEES_ACCOUNT',
  GET_STATUS_ACCOUNT = 'GET_STATUS_ACCOUNT',
  GET_WALLET_IDS_ACCOUNT = 'GET_WALLET_IDS_ACCOUNT',
  DELETE_ACCOUNT = 'DELETE_ACCOUNT',

  CREATE_STATIC_QRCODE_PIX = 'CREATE_STATIC_QRCODE_PIX',

  LIST_TRANSACTION_PIX = 'LIST_TRANSACTION_PIX',
  GET_TRANSACTION_PIX = 'GET_TRANSACTION_PIX',
  CANCEL_TRANSACTION_PIX = 'CANCEL_TRANSACTION_PIX',

  CREATE_SUBSCRIPTION = 'CREATE_SUBSCRIPTION',
  GET_SUBSCRIPTION = 'GET_SUBSCRIPTION',
  GET_PAYMENTS_SUBSCRIPTION = 'GET_PAYMENTS_SUBSCRIPTION',
  LIST_SUBSCRIPTION = 'LIST_SUBSCRIPTION',
  DELETE_SUBSCRIPTION = 'DELETE_SUBSCRIPTION',
  UPDATE_SUBSCRIPTION = 'UPDATE_SUBSCRIPTION',

  CREATE_TRANSFER = 'CREATE_TRANSFER',
  GET_TRANSFER = 'GET_TRANSFER',
  LIST_TRANSFER = 'LIST_TRANSFER',
  CANCEL_TRANSFER = 'CANCEL_TRANSFER',
  CREATE_INTERNAL_TRANSFER = 'CREATE_INTERNAL_TRANSFER',

  CREATE_WEBHOOK = 'CREATE_WEBHOOK',
  GET_WEBHOOK = 'GET_WEBHOOK',
  LIST_WEBHOOK = 'LIST_WEBHOOK',
  DELETE_WEBHOOK = 'DELETE_WEBHOOK',
  UPDATE_WEBHOOK = 'UPDATE_WEBHOOK',
}